requirements.txt
absl-py==0.8.1
alabaster==0.7.12
alembic==1.3.0
argcomplete==1.10.0
argh==0.26.1
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.2
Babel==2.7.0
Beaker==1.10.0
blivet==3.1.6
blivet-gui==2.1.11
Brlapi==0.7.0
cffi==1.12.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.8
decorator==4.4.0
distro==1.4.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-DotEnv==0.1.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
fros==1.1
gast==0.2.2
google-pasta==0.1.8
gpg==1.13.1
grpcio==1.25.0
h5py==2.10.0
humanize==0.5.1
idna==2.8
imagesize==1.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
langtable==0.0.48
lazy-object-proxy==1.4.3
libcomps==0.1.14
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
ntplib==0.3.3
olefile==0.46
opt-einsum==3.1.0
ordered-set==3.1
packaging==19.2
Paste==3.2.2
pathtools==0.1.2
pdoc3==0.7.2
pid==2.2.3
Pillow==7.0.0
ply==3.11
productmd==1.23
pwquality==1.4.2
pycairo==1.18.1
pycodestyle==2.5.0
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.74
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
pyflakes==2.1.1
Pygments==2.5.1
PyGObject==3.34.0
pyinotify==0.9.6
pykickstart==3.21
pylint==2.4.3
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyparted==3.11.2
PyQt5==5.13.1
PyQt5-sip==4.19.19
PySocks==1.7.0
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-meh==0.48
pytz==2019.2
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.2
regex==2019.12.9
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
rpm==4.15.1
selinux==2.9
sepolicy==1.1
setools==4.2.2
simpleline==1.6
six==1.12.0
slip==0.6.4
slip.dbus==0.6.4
snowballstemmer==2.0.0
sos==3.8
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
systemd-python==234
Tempita==0.5.1
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.7
watchdog==0.9.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
